backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
redis==5.2.1
pydantic-settings==2.8.1
email-validator==2.2.0
alembic==1.15.2
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
pytest==8.3.5
httpx==0.28.1
fakeredis==2.28.1
